% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsspy-package.R
\docType{package}
\name{crsspy}
\alias{crsspy}
\alias{crsspy-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/crsspy}
\item BugReports: \url{https://gitlab.com/hrbrmstr/crsspy/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
