% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allows.R
\name{csp_allows}
\alias{csp_allows}
\alias{allows_child_from_source}
\alias{allows_connect_to}
\alias{allows_font_from_source}
\alias{allows_form_action}
\alias{allows_frame_ancestor}
\alias{allows_frame_from_source}
\alias{allows_manifest_from_source}
\alias{allows_media_from_source}
\alias{allows_navigation}
\alias{allows_object_from_source}
\alias{allows_prefetch_from_source}
\alias{allows_script_from_source}
\alias{allows_script_with_nonce}
\alias{allows_style_from_source}
\alias{allows_style_with_nonce}
\alias{allows_unsafe_inline_script}
\alias{allows_unsafe_inline_style}
\alias{allows_worker_from_source}
\title{Tests for what a parsed CSP allows}
\usage{
allows_child_from_source(csp, URL)

allows_connect_to(csp, URL)

allows_font_from_source(csp, URL)

allows_form_action(csp, URL)

allows_frame_ancestor(csp, URL)

allows_frame_from_source(csp, URL)

allows_manifest_from_source(csp, URL)

allows_media_from_source(csp, URL)

allows_navigation(csp, URL)

allows_object_from_source(csp, URL)

allows_prefetch_from_source(csp, URL)

allows_script_from_source(csp, URL)

allows_script_with_nonce(csp, nonce)

allows_style_from_source(csp, URL)

allows_style_with_nonce(csp, nonce)

allows_unsafe_inline_script(csp)

allows_unsafe_inline_style(csp)

allows_worker_from_source(csp, URL)
}
\arguments{
\item{csp}{a \code{csp} object created with \code{\link[=fetch_csp]{fetch_csp()}} or \code{\link[=parse_csp]{parse_csp()}}}

\item{URL}{source or destination URL}

\item{nonce}{a \href{https://html.spec.whatwg.org/multipage/urls-and-fetching.html#attr-nonce}{cryptographic nonce}}
}
\value{
logical
}
\description{
Tests for what a parsed CSP allows
}
\references{
\href{https://www.w3.org/TR/CSP3/}{Content Security Policy Level 3}
}
