% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findings.R
\name{csp_security_checkers}
\alias{csp_security_checkers}
\alias{ensure_csp_df}
\alias{check_script_unsafe_inline}
\alias{check_script_unsafe_eval}
\alias{check_plain_url_schemes}
\alias{check_wildcards}
\alias{check_missing_directives}
\alias{check_ip_source}
\alias{check_deprecated}
\alias{check_nonce_length}
\alias{check_src_http}
\title{Tests for insecure CSP settings}
\usage{
ensure_csp_df(csp_df)

check_script_unsafe_inline(csp_df)

check_script_unsafe_eval(csp_df)

check_plain_url_schemes(csp_df)

check_wildcards(csp_df)

check_missing_directives(csp_df)

check_ip_source(csp_df)

check_deprecated(csp_df)

check_nonce_length(csp_df)

check_src_http(csp_df)
}
\arguments{
\item{csp_df}{Preferably a CSP data frame (made with \code{\link[=as.data.frame]{as.data.frame()}}) but
can be a raw CSP object. Passing in a pre-made data frame will be faster
when using multiple CSP security checker functions.}
}
\value{
a \code{csp_finding} or \code{csp_findings_list} object containing one or more \code{csp_finding}
objects. Each \code{csp_finding} object will have the \code{category}, \code{severity},
\code{message} and \code{where} the violation(s) occurred.
}
\description{
Tests for insecure CSP settings
}
\references{
\href{https://csp.withgoogle.com/docs/index.html}{CSP With Google}
}
