% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-csp-text.R
\name{has_csp}
\alias{has_csp}
\title{Does a URL have a content security policy?}
\usage{
has_csp(URL, method = c("head", "get"))
}
\arguments{
\item{URL}{URL}

\item{method}{method to use fetch CSP (sites may change headers returned
depending on the method used)}
}
\description{
Does a URL have a content security policy?
}
\examples{
has_csp("https://rud.is/")
}
