% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-csp-text.R
\name{parse_csp}
\alias{parse_csp}
\alias{fetch_csp}
\title{Fetch and/or parse a content security policy header value}
\usage{
parse_csp(csp_text, origin_url)

fetch_csp(origin_url, method = c("head", "get"))
}
\arguments{
\item{csp_text}{length 1 character vector containing CSP text}

\item{origin_url}{site to fetch CSP from or to use when just parsing a
plain text (possibly already fetched) CSP}

\item{method}{method to use fetch CSP (sites may change headers returned
depending on the method used)}
}
\description{
Use \code{\link[=fetch_csp]{fetch_csp()}} to load & parse a CSP from a remote site. Use \code{\link[=parse_csp]{parse_csp()}}
to parse an already fetched or composed CSP.
}
\references{
\href{https://www.w3.org/TR/CSP3/}{Content Security Policy Level 3}
}
