% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-csp-text.R
\name{print.csp}
\alias{print.csp}
\title{Printer for CSP objects}
\usage{
\method{print}{csp}(x, ...)
}
\arguments{
\item{x}{CSP object}

\item{...}{unused}
}
\description{
Printer for CSP objects
}
\keyword{internal}
