% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findings.R
\name{print.csp_finding}
\alias{print.csp_finding}
\title{Printer for CSP findings}
\usage{
\method{print}{csp_finding}(x, ...)
}
\arguments{
\item{x}{CSP}

\item{...}{ignored}
}
\description{
Printer for CSP findings
}
\keyword{internal}
