% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{valid_csp_directives}
\alias{valid_csp_directives}
\title{Valid CSP Directives}
\format{An object of class \code{character} of length 27.}
\usage{
valid_csp_directives
}
\description{
A character vector of valid CSP directives as per
Content Security Policy Level 3 W3C Working Draft, 15 October 2018
(\url{https://www.w3.org/TR/CSP3/#iana-registry}).
}
\keyword{datasets}
