% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csp-validate.R
\name{validate_csp}
\alias{validate_csp}
\title{Validate a CSP}
\usage{
validate_csp(csp)
}
\arguments{
\item{csp}{a \code{csp} object created with \code{\link[=fetch_csp]{fetch_csp()}} or \code{\link[=parse_csp]{parse_csp()}}}
}
\value{
data frame
}
\description{
Validates a content security policy. If policy problems are found
a data frame of information notes, warnings and errors is returned
with the specific locations of the issues in the original CSP string.
}
