% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspy-package.R
\docType{package}
\name{crsspy}
\alias{crsspy}
\alias{crsspy-package}
\title{Content Security Policy Decomposer & Evaluator}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/cspy}
\item BugReports: \url{https://gitlab.com/hrbrmstr/cspy/issues}
\item Javadoc: \url{https://www.javadoc.io/doc/com.shapesecurity/salvation/2.6.0}
}
}
\references{
\href{https://www.w3.org/TR/CSP3/}{Content Security Policy Level 3}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
